refactored models.py into a new folder containing DB tables
