+++ models/__init__.py
@@ -1,0 +1,21 @@
+from .base import db
+from .user import User
+from .item import Item
+from .camera_gear import CameraGear
+from .lab_equipment import LabEquipment
+from .location import Location
+from .tag import Tag
+from .associations import item_tags, camera_gear_tags, lab_equipment_tags
+
+__all__ = [
+    'db',
+    'User',
+    'Item',
+    'CameraGear',
+    'LabEquipment',
+    'Location',
+    'Tag',
+    'item_tags',
+    'camera_gear_tags',
+    'lab_equipment_tags',
+]
