+++ models/location.py
@@ -1,0 +1,16 @@
+from .base import db
+from constants import LOCATION_ID, LOCATION_NAME
+
+
+class Location(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(200), nullable=False)
+
+    def __repr__(self):
+        return f"<Location {self.name}>"
+
+    def to_dict(self):
+        return {
+            LOCATION_ID: self.id,
+            LOCATION_NAME: self.name,
+        }
